requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
numpy
pydantic
beautifulsoup4
html2text
tqdm
markdown
markdownify
openai
playwright
python-multipart
setuptools
chromadb
bs4